Move DynamoDB index tests to cover BigTable StorageClient too.
